Move args.py from scripts/ to scr/azul/ (#3249)
